fefactor: db 操作を controller に移動
